cleaned up `make clean`

+++ Makefile
@@ -2,8 +2,9 @@
 SHELL := /bin/bash
 HOST ?= 127.0.0.1
 PORT ?= 3000
+BROWSER ?= firefox
 export HOST PORT
-all: run
+all: run view
 getting-started: Dockerfile Makefile
 	docker build -t $@ $(<D)
 	touch $@
@@ -11,12 +12,14 @@ getting-started: Dockerfile Makefile
 	envsubst < $< > $@
 run: getting-started
 	docker run --detach --publish $(HOST):$(PORT):$(PORT) $< > $<
+view:
+	$(BROWSER) $(HOST):$(PORT)/
 stop:
-	-docker stop $$(<getting-started)
+	if [ -s "getting-started" ]; then docker stop $$(<getting-started); fi
 clean:
 	$(MAKE) stop
-	-docker rm $$(<getting-started)
-	-docker rmi getting-started
+	if [ -s "getting-started" ]; then docker rm $$(<getting-started); fi
+	if [ -f "getting-started" ]; then docker rmi getting-started; fi
 	rm -f getting-started
 distclean: clean
 	rm -f Dockerfile
